requirements.txt
flask
flask-jwt-extended
flask-marshmallow
flask-restful
flask-sqlalchemy
flask-limiter
flask-login
flask-admin
python-dotenv==0.14.0
marshmallow-enum
marshmallow-sqlalchemy
